More sensible default en-US locale.
The previous default en-US locale used 24-hour time. This matched the default
behavior of the `date` command on macOS, but isn’t the local convention. These
new definitions are based on the behavior of date.toLocale{,Date,Time}String in
Chrome, Node and Firefox. (Safari is a little different.)

    %C: "%a %b %e %X %Y" ↦ "%x, %X"
    %x: "%H:%M:%S" ↦ "%-I:%M:%S %p"
    %X: "%m/%d/%Y" ↦ "%-m/%-d/%Y"

Fixes #19.